+++ src/cl/loccet/model/Proyecto.java
@@ -1,14 +1,44 @@
 package cl.loccet.model;
 
-public class Proyecto {
+import java.util.ArrayList;
+import java.util.HashMap;
 
-    private int id;
+public class Obra {
+    private Integer id;
+    private String direccion;
+    private ArrayList<cl.loccet.model.Trabajador> listaEmpleados;
+    private HashMap<String, cl.loccet.model.Trabajador> mapEmpleados;
+    //TODO: Fecha inicio y fin.
 
-    public Proyecto(int id) {
+    public Obra(int id, String ubicacion){
+        this.id = new Integer(id);
+        this.direccion = ubicacion;
+        listaEmpleados = new ArrayList<>();
+        mapEmpleados = new HashMap<>();
+    }
+
+    //Setter
+    public void setId(int id){
         this.id = id;
     }
 
-    public int getId() {
+    public void setUbicacion(String ubicacion) {
+        this.direccion = ubicacion;
+    }
+
+    //Getter
+    public Integer getId() {
         return id;
+    }
+
+    public String getUbicacion() {
+        return direccion;
+    }
+
+    //Metodos
+
+    public void agregarTrabajador(cl.loccet.model.Trabajador trabajador){
+        listaEmpleados.add(trabajador);
+        mapEmpleados.put(trabajador.getRut(), trabajador); //TODO: Validar el put
     }
 }
